llm_service/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
google-generativeai==0.8.4
python-dotenv==1.0.1
